fix: 健壮性批次三——数值/配置判型修复——Boss json 直读路径元素级判型补齐（strafe_speeds 坏元素/非正值钳下限 1.0——原 (float)v.AsDouble() 直转坏元素静默为 0 使对应模式走位彻底静止且无运行期兜底；fire_intervals 非数值元素回退默认对应位——防难度缩放把 0 写回缓存，运行期另有 AB3 IntervalFloor 兜底）；HealthChanged/DifficultyChanged 信号声明 double→float（U16「参数统一 float」全量落地——原声明 double 而监听端 Callable.From<float> 依赖 Variant 隐式窄化，发射中转处显式窄化一次，监听端零改动）；MissionsService.RefreshMissions 的 goal 与 InitMissions 同口径转 int（原 Variant 原样入典同键双类型并存）；RunProgressionService.ApplyProgressionParams 本地钳 timeStepSeconds ≥0.1（两处 RunTime/step 除法不再依赖上游 GameState 注入点钳制）。核实不动：Spawner/Enemy 机型 config 直接索引为上游双重保证下的 fail-fast 契约——MergeType 对 hp/speed/score/fire 等已做元素级判型（L05/AB8），strategies 键不经 json 覆盖路径恒为内置表 typed 非空数组，本地硬转型在契约破坏时崩比静默回退（需瞎编默认数值）更合理；Main 子弹时间恢复 ramp 的缩放帧长隐性放大确认为演出依赖，补注释陈述约束。构建零警告+导入无警告+主场景 300 帧无错误

+++ csharp/godot/GameState.cs
@@ -25,10 +25,10 @@ public partial class GameState : Node
     public delegate void ComboChangedEventHandler(int combo);
 
     [Signal]
-    public delegate void HealthChangedEventHandler(double newHealth);
+    public delegate void HealthChangedEventHandler(float newHealth);
 
     [Signal]
-    public delegate void DifficultyChangedEventHandler(double newMultiplier);
+    public delegate void DifficultyChangedEventHandler(float newMultiplier);
 
     [Signal]
     public delegate void DifficultySelectedEventHandler(StringName difficulty);
@@ -379,7 +379,7 @@ public partial class GameState : Node
     // （HealthChanged/AugmentsChanged；触发点均为运行期对局事件/玩家操作——LoseHealth/Heal/AddBuff/
     // ConsumeAugment，晚于 _Ready 本订阅；ResetRun/天赋路线（TalentService 层级写入）直发路径
     // 不经本事件，订阅重发不与之重复）
-    private void OnCombatHealthChanged(double v) => EmitSignal(SignalName.HealthChanged, v);
+    private void OnCombatHealthChanged(double v) => EmitSignal(SignalName.HealthChanged, (float)v);
 
     private void OnCombatAugmentsChanged() => EmitSignal(SignalName.AugmentsChanged);
 
@@ -415,7 +415,7 @@ public partial class GameState : Node
     // 对局进程域（第五轮拆域）：RunProgressionService C# 事件 → GameState 同名信号转发
     // （DifficultyChanged/DifficultySelected；触发点均为运行期对局事件/玩家操作——_Process
     // 时间档重算/SetDifficulty，晚于 _Ready 本订阅；AddBossKill 直发路径不重复）
-    private void OnRunProgDifficultyChanged(double v) => EmitSignal(SignalName.DifficultyChanged, v);
+    private void OnRunProgDifficultyChanged(double v) => EmitSignal(SignalName.DifficultyChanged, (float)v);
 
     private void OnRunProgDifficultySelected(StringName v) => EmitSignal(SignalName.DifficultySelected, v);
